refactor: extract module payload types to dedicated file
Move RuleModulePayloadMap and composite payload interfaces
(MagicModulePayload, LifestyleModulePayload, SkillsModulePayload,
ProgramsModulePayload) from loader-types.ts to module-payloads.ts.

Re-exports from loader-types.ts preserve backward compatibility.
Reduces loader-types.ts from 1,732 to 1,600 lines.
